fix: only add layer surface after first configure
in the wlr layer surface protocol, you are only allowed to attach a
buffer after the first configure from the compositor. this was not
respected here and would lead to a race-condition where most of the time
the layer surface creation was successful (when the compositor was fast
enough) and on other times it would raise a protocol error

disclaimer: I have absolutely no clue what I am doing here, as I don't fully grasp
the inner event loop of iced. but it seems to work so far on my end